Advanced_RowID implemented and AdvancedTable.get added. Refactored Table Tests
