+++ alcustoms.pyproj
@@ -132,6 +132,9 @@
     <Compile Include="sql\objects\View.py" />
     <Compile Include="sql\objects\__init__.py" />
     <Compile Include="sql\tests\objects\test_graphdb.py" />
+    <Compile Include="sql\tests\objects\test_Table.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="sql\tests\objects\test_utilities.py">
       <SubType>Code</SubType>
     </Compile>
